requirements.txt
pygame
numpy
pillow
scipy
varint